obstacles trigger sound
