+++ Assets/Audio/479672__craigsmith__r12-60-horses-snort-and-vocalize copy 2.wav.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 69918f7751d604b0d8b76d91cd80f1a2
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 7
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
