+++ Assets/Prefabs/Creatures/Comet Horse.prefab
@@ -1,0 +1,71 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &3326943634093517293
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1654313585272908052, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_Name
+      value: Comet Horse
+      objectReference: {fileID: 0}
+    - target: {fileID: 2152502805223108739, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_Sprite
+      value: 
+      objectReference: {fileID: 21300000, guid: 2f684f4af8a0845e2ba59a5718f95687, type: 3}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8574308485787647591, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 9116801771802529537, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: smooth
+      value: 50
+      objectReference: {fileID: 0}
+    - target: {fileID: 9116801771802529537, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
+      propertyPath: moveSpeed
+      value: 10
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 8e3e6906df6064bb380715f10833d4f7, type: 3}
