+++ Assets/Prefabs/Creatures/Comet Horse.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8685446086ff24bac8a39f976c7e5061
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
