+++ Assets/Prefabs/HorseMovementHandler.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b264f92d64e0c474abeeb09429ea9cbb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
