+++ Assets/Prefabs/Obstacle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 52af3025c71254846826307dd7acdfd3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
