--- Assets/Prefabs/PlayerInputHandler.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9e0c5b2ccbf08497a9336d33569adead
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
