+++ Assets/Scripts/HorseMovementHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09c6eb64bb1a947d8800d6f28c0be6d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
