+++ Assets/Scripts/Obstacle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 625996cc2f6584a1896963abff02cb09
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
