--- Assets/Scripts/PlayerMovementHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2a052dc3b109d4805906463d11150a9b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
